Fix for uncatchable `AbortError`s
Patching `window.fetch` without handling the `AbortError` results in
JS code that uses fetch, becoming unable to catch these errors. As such,
application errors will be emitted to the console (and in the case of
webpack-dev-server, to the browser window), that are very mysterious.
Even if one has try/catch blocks within an async function, or then/catch
blocks for the Promise using `fetch()`, these errors will not be
catched.

I am not fully in the know as to why it works this way (the internals of
an async/Promise are very convoluted especially when `this` is involved
as in here), but I suspect the patched call will emit its own
`AbortError` extra to the one that comes from the original one, and the
uncatched one in the patched function will resurface even after catching
one of them.